Attempt to refactor initialization
