+++ clients/fides-js/src/lib/initialize.ts
@@ -1,0 +1,271 @@
+import { gtm } from "../integrations/gtm";
+import { meta } from "../integrations/meta";
+import { shopify } from "../integrations/shopify";
+import { getConsentContext } from "./consent-context";
+import { initOverlay } from "./consent";
+import {
+  CookieKeyConsent,
+  CookieIdentity,
+  CookieMeta,
+  getOrMakeFidesCookie,
+  makeConsentDefaultsLegacy,
+  buildCookieConsentForExperiences,
+  FidesCookie,
+  isNewFidesCookie,
+} from "./cookie";
+import {
+  PrivacyExperience,
+  FidesConfig,
+  FidesOptions,
+  UserGeolocation,
+  ConsentMethod,
+  SaveConsentPreference,
+  ConsentMechanism,
+} from "./consent-types";
+import {
+  constructFidesRegionString,
+  debugLog,
+  experienceIsValid,
+  transformConsentToFidesUserPreference,
+  validateOptions,
+} from "./consent-utils";
+import { fetchExperience } from "../services/fides/api";
+import { getGeolocation } from "../services/external/geolocation";
+import { OverlayProps } from "../components/Overlay";
+import { updateConsentPreferences } from "./preferences";
+import { resolveConsentValue } from "./consent-value";
+
+export type Fides = {
+  consent: CookieKeyConsent;
+  experience?: PrivacyExperience;
+  geolocation?: UserGeolocation;
+  options: FidesOptions;
+  fides_meta: CookieMeta;
+  gtm: typeof gtm;
+  identity: CookieIdentity;
+  init: (config: FidesConfig) => Promise<void>;
+  initialized: boolean;
+  meta: typeof meta;
+  shopify: typeof shopify;
+};
+
+declare global {
+  interface Window {
+    Fides: Fides;
+  }
+}
+
+const retrieveEffectiveRegionString = async (
+  geolocation: UserGeolocation | undefined,
+  options: FidesOptions
+) => {
+  // Prefer the provided geolocation if available and valid; otherwise, fallback to automatically
+  // geolocating the user by calling the geolocation API
+  const fidesRegionString = constructFidesRegionString(geolocation);
+  if (!fidesRegionString) {
+    // we always need a region str so that we can PATCH privacy preferences to Fides Api
+    return constructFidesRegionString(
+      // Call the geolocation API
+      await getGeolocation(
+        options.isGeolocationEnabled,
+        options.geolocationApiUrl,
+        options.debug
+      )
+    );
+  }
+  return fidesRegionString;
+};
+
+const automaticallyApplyGPCPreferences = (
+  cookie: FidesCookie,
+  fidesRegionString: string | null,
+  fidesApiUrl: string,
+  effectiveExperience?: PrivacyExperience | null
+) => {
+  if (!effectiveExperience || !effectiveExperience.privacy_notices) {
+    return;
+  }
+
+  const context = getConsentContext();
+  if (!context.globalPrivacyControl) {
+    return;
+  }
+
+  let gpcApplied = false;
+  const consentPreferencesToSave = effectiveExperience.privacy_notices.map(
+    (notice) => {
+      if (
+        notice.has_gpc_flag &&
+        !notice.current_preference &&
+        notice.consent_mechanism !== ConsentMechanism.NOTICE_ONLY
+      ) {
+        gpcApplied = true;
+        return new SaveConsentPreference(
+          notice,
+          transformConsentToFidesUserPreference(false, notice.consent_mechanism)
+        );
+      }
+      return new SaveConsentPreference(
+        notice,
+        transformConsentToFidesUserPreference(
+          resolveConsentValue(notice, context),
+          notice.consent_mechanism
+        )
+      );
+    }
+  );
+
+  if (gpcApplied) {
+    updateConsentPreferences({
+      consentPreferencesToSave,
+      experienceId: effectiveExperience.id,
+      fidesApiUrl,
+      consentMethod: ConsentMethod.gpc,
+      userLocationString: fidesRegionString || undefined,
+      cookie,
+    });
+  }
+};
+
+/**
+ * Get the initial Fides cookie based on legacy consent values
+ * as well as any preferences stored in existing cookies
+ */
+export const getInitialCookie = ({ consent, options }: FidesConfig) => {
+  // Configure the default legacy consent values
+  const context = getConsentContext();
+  const consentDefaults = makeConsentDefaultsLegacy(
+    consent,
+    context,
+    options.debug
+  );
+
+  // Load any existing user preferences from the browser cookie
+  const cookie: FidesCookie = getOrMakeFidesCookie(
+    consentDefaults,
+    options.debug
+  );
+  return cookie;
+};
+
+/**
+ * If saved preferences are detected, immediately initialize from local cache
+ */
+export const getInitialFides = ({
+  cookie,
+  experience,
+  geolocation,
+  options,
+}: {
+  cookie: FidesCookie;
+} & FidesConfig) => {
+  const initialFides: Partial<Fides> = {};
+  const hasExistingCookie = !isNewFidesCookie(cookie);
+  if (!hasExistingCookie) {
+    return null;
+  }
+
+  initialFides.consent = cookie.consent;
+  initialFides.fides_meta = cookie.fides_meta;
+  initialFides.identity = cookie.identity;
+  initialFides.experience = experience;
+  initialFides.geolocation = geolocation;
+  initialFides.options = options;
+  initialFides.initialized = true;
+  return initialFides;
+};
+
+/**
+ * The bulk of the initialization logic
+ * 1. Validates options
+ * 2. Retrieves geolocation
+ * 3. Retrieves experience
+ * 4. Updates cookie
+ * 5. Initialize overlay components
+ * 6. Apply GPC if necessary
+ */
+export const initialize = async ({
+  cookie,
+  options,
+  experience,
+  geolocation,
+}: { cookie: FidesCookie } & FidesConfig) => {
+  const updatedFides: Partial<Fides> = {};
+  const context = getConsentContext();
+
+  let shouldInitOverlay: boolean = options.isOverlayEnabled;
+  let effectiveExperience: PrivacyExperience | undefined | null = experience;
+  let fidesRegionString: string | null = null;
+
+  if (shouldInitOverlay) {
+    if (!validateOptions(options)) {
+      debugLog(
+        options.debug,
+        "Invalid overlay options. Skipping overlay initialization.",
+        options
+      );
+      shouldInitOverlay = false;
+    }
+
+    fidesRegionString = await retrieveEffectiveRegionString(
+      geolocation,
+      options
+    );
+
+    if (!fidesRegionString) {
+      debugLog(
+        options.debug,
+        `User location could not be obtained. Skipping overlay initialization.`
+      );
+      shouldInitOverlay = false;
+    } else if (!effectiveExperience) {
+      effectiveExperience = await fetchExperience(
+        fidesRegionString,
+        options.fidesApiUrl,
+        cookie.identity.fides_user_device_id,
+        options.debug
+      );
+    }
+
+    if (
+      effectiveExperience &&
+      experienceIsValid(effectiveExperience, options)
+    ) {
+      // Overwrite cookie consent with experience-based consent values
+      // eslint-disable-next-line no-param-reassign
+      cookie.consent = buildCookieConsentForExperiences(
+        effectiveExperience,
+        context,
+        options.debug
+      );
+
+      if (shouldInitOverlay) {
+        await initOverlay(<OverlayProps>{
+          experience: effectiveExperience,
+          fidesRegionString,
+          cookie,
+          options,
+        }).catch(() => {});
+      }
+    }
+  }
+  if (shouldInitOverlay) {
+    automaticallyApplyGPCPreferences(
+      cookie,
+      fidesRegionString,
+      options.fidesApiUrl,
+      effectiveExperience
+    );
+  }
+
+  // Initialize the window.Fides object
+  updatedFides.consent = cookie.consent;
+  updatedFides.fides_meta = cookie.fides_meta;
+  updatedFides.identity = cookie.identity;
+  updatedFides.experience = experience;
+  updatedFides.geolocation = geolocation;
+  updatedFides.options = options;
+  updatedFides.initialized = true;
+
+  return updatedFides;
+};
